+++ client/app/components/Pages/Polling.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import "whatwg-fetch";
 import Header from "../Header/Header";
 import Polling from "../Polling/Polling";
-import Map from "../Map/Map";
+
 
 import { getFromStorage, setInStorage } from "../../utils/storage";
 
@@ -22,7 +22,7 @@ class Ballot extends Component {
       <div>
         <Header />
         <div className="jumbotron row" id="pollingJumbo">
-          <Map />
+  
           <Polling />
         </div>
       </div>
